moving addcompostieDV to DVGeoSketch
